Added logic to traverse to device group from iotsupervisor
Added logic to traverse from supervisor to iot group.  Added external call to child device to add a temperature.
I'm not sure if this works.  Will have to play with it.  Still needing to get the temperatures from all the devices.
This, so far, is very imperative inside the IotMain, but right now, just looking to see if the plumbing works.