Added all logic, but problems in plumbing between device groups and device manager
I've added logging to the calls.  The call IotDeviceManager.RequestAllGroupTemperatures makes it all the way down to the device (although the request id seems different), but then fails to come back up the ladder, causing the whole system to timeout.  Need to investigate how to get this to traverse upward.  I believe if I were to call the device group directly, then I would be able to make this happen on the group level.  This may be the next step needed in testing this.  Let's see if that works correctly, before tackling the issue between device group and device manager.