Got device-group respondAllTemperatures to work by itself.
I got the IotDeviceGroup.RequestAllTemperatures to come back with a valid response.
In order to do this, I got a reference to a IotDevice and added a temp directly, using
the paths showing up in the log messages.  Originally, I had the wrong paths setup.  However,
even with this, it behaved and returned a proper message.  I still need to get the IotDeviceManager.RequestAllGroupTemperatures to work properly.  It just times out.