Got the IotDeviceManager.RequestAllGroupTemperatures to work properly.
Figured out the problem with the timeouts for RequestAllGroupTemperatures.  When creating 'Query' Actors (second-class citizens, for this purpose), the response message coming back from the Command-Actor
was listes as originating from the Command-Actor, rather than the Actor that created it.  This caused a mismatch between the original groupToIdMap key values.  Setting the originating actor to the parent
of the 'Query' actor fixed issues.